+++ src/Storage/Storage.Management/ChangeLog.md
@@ -18,6 +18,17 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Supported NFS File Share and NFS file and directory properties
+    - `New-AzStorageShare`
+    - `Get-AzStorageShare`
+    - `Get-AzStorageFile`
+    - `Get-AzStorageFileContent`
+    - `New-AzStorageDirectory`
+    - `Remove-AzStorageFile`
+    - `Set-AzStorageFileContent`
+    - `Start-AzStorageFileCopy`
+* Supported create hard link in NFS File Share 
+    - `New-AzStorageFileHardLink`
 
 ## Version 8.2.0
 * Supported new SkuName when create/update Storage account for Files Provisioned v2 account type:  "StandardV2_LRS", "StandardV2_GRS", "StandardV2_ZRS", "StandardV2_GZRS", "PremiumV2_LRS", "PremiumV2_ZRS"